build.sh with --clean option

+++ build.sh
@@ -1,11 +1,30 @@
+# only one argument allowed
+if [ $# -eq 1 ] && [ "$1" != "--clean" ]; then
+    echo "usage: build.sh [--clean]"
+    exit 1
+fi
 
+# get the CLEAN flags
+if [ "$1" = "--clean" ]; then
+    INSTALL_FLAG="-f"
+    BUILD_FLAG=""
+else
+    INSTALL_FLAG=""
+    BUILD_FLAG="--skip-existing"
+fi
+
+# get the latest changes from github
 git reset --hard
 git checkout nersc
 git pull origin nersc
 
+# get the bundle-anaconda command
 source /usr/common/contrib/bccp/python-mpi-bcast/activate.sh
+
+# activate our root anaconda install to start
 source /usr/common/contrib/bccp/anaconda3/bin/activate root
 
+# make the recipes
 python extrude_recipes requirements.yml
 
 build ()
@@ -13,17 +32,20 @@ build ()
     local PYTHON=$1
     
     pushd recipes
-    conda build --no-test --python $PYTHON mpi4py
+    conda build --python $PYTHON mpi4py
     for f in *; do
-        conda build --no-test --python $PYTHON $f
+        if [ $f != "mpi4py" ]; then
+            conda build --python $PYTHON $BUILD_FLAG $f
+        fi
     done
     
+    # install packages into this python version's environment
     source activate $PYTHON
-    
-    conda install --use-local -f --yes *
+    conda install $INSTALL_FLAG --use-local --yes *
     conda update --use-local --yes *
     popd
 
+    # and tar the install
     bundle-anaconda /usr/common/contrib/bccp/anaconda3/envs/bccp-anaconda-$PYTHON.tar.gz $CONDA_PREFIX
     
 }
